Fix timestamp and one hour floor logic
One hour floor will now be initialized to '---' to indicate absence of
the floor. Additionally, a check is done to see if the value of the
floor is real before setting the state.

Timestamps will now be retrieved on component mount instead of waiting
for the interval.
The sales timestamps are checked if they are greater than or equal to
the retrieved timestamps. This ensure that a sale right on the dot is
included in the count.