Add checks to retrieved timestamps
This change ensures that the previous useEffect, which calculates
the timestamps of tenMinutesAgo, oneHourAgo, and oneDayAgo, is complete
before starting to calculate number of sales and one hour floor. This
fixes the bug of the one hour floor being calculated even if there
were no sales in the past hour.