Fix 1H floor showing previous collection 1H floor
If a user is initially on a collection page that had no sales within
the past hour, the floor will show correctly. If they navigate to a
collection that does have a sale in the past hour and then navigate
back, they will see the initial collection will have a 1H floor that
belongs to the active collection.

This was because the 1H floor was 99999 upon navigating back to the
inactive collection, which does not set the 1H floor state and
therefore keeps the state of the active collection. Adding the else
statement will set the 1H floor state to '---' every time there has
been no sales in the past hour.